[feat] 에러 팝업 생성
