+++ design/src/main/res/drawable/ic_page_error.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="50dp"
+    android:height="50dp"
+    android:viewportWidth="50"
+    android:viewportHeight="50">
+  <path
+      android:pathData="M24.999,36.993C25.394,36.993 25.726,36.859 25.993,36.592C26.26,36.325 26.394,35.993 26.394,35.598C26.394,35.203 26.26,34.872 25.993,34.605C25.726,34.337 25.394,34.204 24.999,34.204C24.604,34.204 24.273,34.337 24.005,34.605C23.738,34.872 23.604,35.203 23.604,35.598C23.604,35.993 23.738,36.325 24.005,36.592C24.273,36.859 24.604,36.993 24.999,36.993ZM24.023,28.346H25.975V11.611H24.023V28.346ZM25.011,49.266C21.656,49.266 18.501,48.629 15.546,47.356C12.592,46.083 10.021,44.354 7.836,42.171C5.65,39.988 3.92,37.421 2.645,34.471C1.37,31.52 0.732,28.366 0.732,25.011C0.732,21.656 1.369,18.501 2.642,15.546C3.916,12.592 5.644,10.021 7.827,7.836C10.01,5.65 12.577,3.92 15.527,2.645C18.479,1.37 21.632,0.732 24.987,0.732C28.342,0.732 31.497,1.369 34.452,2.642C37.407,3.916 39.977,5.644 42.162,7.827C44.348,10.01 46.078,12.577 47.353,15.527C48.628,18.479 49.266,21.632 49.266,24.987C49.266,28.342 48.629,31.497 47.356,34.452C46.083,37.407 44.354,39.977 42.171,42.162C39.988,44.348 37.421,46.078 34.471,47.353C31.52,48.628 28.366,49.266 25.011,49.266ZM24.999,47.313C31.229,47.313 36.505,45.152 40.828,40.828C45.152,36.505 47.313,31.229 47.313,24.999C47.313,18.77 45.152,13.493 40.828,9.17C36.505,4.847 31.229,2.685 24.999,2.685C18.77,2.685 13.493,4.847 9.17,9.17C4.847,13.493 2.685,18.77 2.685,24.999C2.685,31.229 4.847,36.505 9.17,40.828C13.493,45.152 18.77,47.313 24.999,47.313Z"
+      android:fillColor="#D8D8D8"/>
+</vector>
